requirements.txt
# Django Core
Django==5.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# CORS
django-cors-headers==4.3.1

# Rate Limiting
django-ratelimit==4.1.0

# HTTP Requests
requests==2.31.0

# Environment
python-decouple==3.8

# Utilities
python-dateutil==2.8.2

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0
